changing more table parts

+++ bamazon.sql
@@ -3,7 +3,7 @@ CREATE DATABASE bamazon_db;
 use bamazon_db;
 
 CREATE TABLE products (
-	 item_id INTEGER(10) AUTO_INCREMENT NOT NULL,
+	 item_id INT AUTO_INCREMENT NOT NULL,
 
 	 product_name VARCHAR(20) NOT NULL,
